Update patch set 2
Thanks for all the suggestions. I was hoping that it wouldn't be too painfully obvious that I haven't written a line of Java in at least 8 years. Oh well. :)

I made all those changes, and added some documentation. I'm not sure if I captured the output of the hook in the best way.

Patch-set: 2