Update patch set 4
To improve the power of the ACL system I'm starting to look at http://jboss.org/drools as a rule engine we could embed into Gerrit, allowing higher-level rules to be written by the site administrator and/or project owners.  As I'm thinking about this in conjunction with all the hooks you mention, we might be able to take your EventQueue execution code and support hooks by allowing the administrator/project owner to schedule execution when a rule predicate matches, and sprinkle enough calls to Drools around in the code that pretty much everything can be captured.

I only started looking at this last night, so I haven't even had a chance to attempt a prototype integration of Drools.  But I do know we want to inject rules on Linux kernel forks that say "commit must have signed-off-by line for author".  We also want to be able to say "CC android-merge-log on all submitted changes under platform/*".  It wouldd be nice if these could (for the most part) run purely in Java, without the overhead of an external hook invocation.

So in theory, we could also have rules like "when a patch set is replaced, invoke script $site_dir/replaced-patch".

I dunno.  Its still pie-in-the-sky.

But I do agree with you in principle, we want all of these events to be something an administrator can hook logic onto and at least receive notification of them.

Patch-set: 4