Update patch set 1
Patch Set 1: Verified

I'm wondering if just silently ignoring invalid characters is a good idea from user interface point of view.

Is it difficult to make the background of input box to blink as red (probably fading out) when user types disallowed characters?

It may seem that I'm asking for unnecessary fanciness but that sort of things make UI pleasant to use. :)

Patch-set: 1