Update patch set 5
Patch Set 5: No score

Would it be possible to merge this patch while a better rule based system is being worked on? Perhaps the feature should be turned off by default forcing the administrator to actively set the flag and in this way minimizing impact when the flag is removed in favor of a better system.

Patch-set: 5
Reviewer: Gerrit User 1001814 <1001814@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=0