Update patch set 5
Sounds great! I have not looked in detail at the code but right now it seems to me that the code will be merged whatever the hook-script returns. Am I right? Is it possible at this stage in the code to call Reject() if the hook returns an error return-value?
We are verifying content of the commit-message and would prefer to do this on the server side, that is why I ask.

Patch-set: 5