Update patch set 1
OK, I think I get the problem.

During authenticate(), we open two DirContext connections.  The first one is anonymous, which fails, and the whole thing aborts horribly.  So that needs to change to use the supplied username/password if it was presented.  So the logic of lines 314-319 needs to be hoisted up a bit earlier to line 310, and its DirContext needs to be reused:

        if (authConfig.getAuthType() == AuthType.LDAP) {
          // We found the user account, but we need to verify
          // the password matches it before we can continue.
          //
          authenticate(m.getDN(), who.getPassword());
        }

This seems like it should be simple to fix.  Modify the code to call authenticate() here instead of open() at line 310 when authType is LDAP.  Modify authenticate() to return the DirContext instead of just opening and discarding it.  Use that returned DirContext for the rest of that method.

During groups(), lookup() and lookupGroups() if we don't have the data we need in cache, or the cache is stale, we again are opening an anonymous DirContext to acquire the data.  Which will fail.  So these may need to login using the calling user's username/password... which we don't have anymore because these requests are being made in subsequent web requests.

To fix this, we're probably going to have to store in memory the user's username and password in a cache, so we can pull them out and reuse them to open a subsequent connection to load the group cache, or to lookup groups by name when configuring a group in the admin panel, etc.

This is much more difficult.  What you need here is a Provider<CurrentUser> to tell you who the caller of your method is, but right now you can't get one because there are two different variants (one for SSH, and another for HTTP) and this is the generic server module where the specific variant isn't known.  Ick.  You might be able to develop a CommonCurrentUserProvider which actually has two Provider<CurrentUser> objects inside of it for each connection type, and just tests them sequentially until it gets a hit.

And then take an injection here for CommonCurrentUserProvider, and ask it who the current user is.  Once you find the user, look in the cache for their username/password.  And obviously, during authenticate(), we'd need to save their username/password into the cache, keyed by their Account.Id, so we can get it later via that CommonCurrentUserProvider.

I'd suggest doing this as two commits.  The simple thing is to fix authenticate() to use the proper DirContext on the initial bind in authenticate().  Then do a second commit that attempts to fix the other methods in this class to use the cached username/password information.

Patch-set: 1