Update patch set 1
Patch Set 1: No score

Shawn,

I replaced the open call with authenicate and got it all to work, but wasn't sure if you wanted to get rid of the separate two-stage bind and then authenticate behavior. 

I was going to keep the 'this.username' check and if it is set (the username being set in the gerrit.config file) then bind to it and then use the dialog box settings to authenticate against. Otherwise, if 'this.username' is null, then just bind and authenticate all together with the dialog box settings.

The way I have it coded now, it just ignores whenever is set in the gerrit.config file, and always binds and authenticates based on what the user has typed into the dialog box.

Thoughts?

By the way, should I continue using this bug to chat back and forth on this, or should I be using something else (email, im, etc.)

Patch-set: 1
Label: Code-Review=0