Update patch set 4
Patch Set 4: No score; Looks good to me, but someone else must approve

So I think this gets you the authentication mode you needed.  But there's at least one other problem in the class, aside from the group lookup needing a cached username/password.

During construction of LdapRealm we invoke discoverLdapType() which is trying to figure out if this is an OpenLDAP server or a Microsoft Active Directory Server.  Except its doing so by creating an anonymous connection.  The query patterns we need for the servers are different, and we are supposed to automatically discover them on startup, but we can't, because we can't login to your directory server without an actual end-user's identity.

So we need to refactor these fields to be lazily initialized on the first authentication attempt.

Patch-set: 4