Update patch set 1
Patch Set 1: No score

Logic looks sound.  A few nits related to copy and paste problems that we should fix before we submit this code.  Also, it would be nice if we could refactor this with SubmitAction because I think some of these bugs actually exist in SubmitAction... :-(

Patch-set: 1
Label: Verified=0