Update patch set 1
Patch Set 1:

But without an href= attribute, the browser shouldn't render them as clickable blue underlined text. That is, <a>foo</a> (which is what you have written with the code you have) shouldn't be special in any way.

Having an href target makes an anchor special in a clickable way. Having a name attribute makes an anchor special in that it a browser might want to offer this anchor in a table-of-contents or outline view (which no browser does, sadly).

Having neither makes it a pointless element. And a browser that makes it clickable is sort of wrong. If Android is wrong here, well, IMHO Android is wrong and we shouldn't be relying on their wrong-ness to get the effect you want.

If you really want it look clickable, it needs an href, or a CSS class applied to it that makes it look clickable.

Patch-set: 1