Update patch set 4
Patch Set 4:

The code looks good now I think, but I have some concerns with the feature approach that I feel I should at least express:

1) I am not convinced that this if going to be very helpful to you or anybody, and 2) I am leery of treading down this path with emails just yet.

Let me explain:

1) Most reviewers are added to changes after the first patch is submitted to Gerrit.  So, I believe that this will likely only prevent emails from going to those reviewers added on the command line during the change submission.

2) While I have had many complaints from our users about receiving too many Gerrit emails, I am leery of treading down the path of giving users fine grained control over turning emails on and off.  One reason is that often times when the users complain, they have valid reasons and it indicates something that should be fixed for everyone in Gerrit.  If we allow users to simply turn off a specific email, it could have some undesired consequences:

A) It may cause users to simply turn the bad emails off and no longer complain to us about it, preventing us from improving the general workflow of Gerrit for everyone.


B) It may interfere with the general designed assumptions built into the Gerrit workflow.  Much of the workflow is designed around the assumption that users will receive an appropriate notification at appropriate points in the workflow so that users of Gerrit do not have to go outside of Gerrit to contact other users.  If users suddenly realise that some users may be turning off certain automated emails from Gerrit, they may simply start sending emails to those users externally to Gerrit to ensure that they are notified.  This would mean that our attempt to reduce SPAM may in fact increase it since now users may get the Gerrit emails (if they have not turned them off) along with the external emails. 

So, my preference for now is to try to improve the workflow as much as possible and reduce the emails for everyone with better approaches in the main Gerrit workflow.  I think that we are regularly making progress here and I feel that several people have had some very good recent suggestions where it can still be improved.  So, until those suggestions are exhausted, I would prefer to avoid this fine grained approach.

If turning emails off is still a strong desire, I could more easily support the idea that users should be able to turn Gerrit emails off entirely for themselves.  But to support this, I would suggest one other important improvement in Gerrit along with this:  that these users be treated differently in Gerrit.  Perhaps these users should appear differently in the UIs so that when they are added as reviewers, it is obvious that they are squelching emails and thus will not receive the normal notifications built into Gerrit.  This way, for other users one can assume that they will receive the normal Gerrit notifications and people will likely not start regularly emailing around Gerrit.

Patch-set: 4