Update patch set 4
Patch Set 4:

I agree with all your points especially "B", rethinking this again, I believe this will confuse users with the workflow of Gerrit. Improving the workflow will be a culture shift, and I agree it needs to be thought out better with the community in mind.

How about this, if we make this a project setting, then it will not interfere with the workflow. For now, users will not be able to change it, only projects owners can.

That will allow different workflows per project. Projects that enable the flag may send emails on every action or emails on explicit submission (via Publish comments). This will allow Gerrit to imitate Rietveld workflow.

Later on, we could implement some sort advanced email setup. But in the short term, this will help solve the massive load of emails that everyone is getting. They can wish to set it up per project.

What do you say?

Patch-set: 4