Update patch set 1
Patch Set 1: Verified; Looks good to me, approved

I think this is what you folks meant about caching the shortest example strings in RefControl. It isn't the production of the shortest example string that is the problem, its the need to run the sort routine on every access request.

I'm not bothering to cache the shortest examples because after all of these changes are in the shorted example is only required when access control rules are viewed/modified, or when there are changes and a new sorting must be computed by the cache.

If computing the shortest examples for the admin access control panel is still a a problem, we can consider adding another cache there... but I'm hoping we can avoid it.

Patch-set: 1
Reviewer: Gerrit User 1001240 <1001240@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+2
Label: Verified=+1