Update patch set 6
Patch Set 6:

I wanted to write more, but here is a quick summary of my thoughts.  This codes looks good, but it is not what I had in mind in my proposal, it simply squelches emails, but does not actually provide a publishing mechanism for patches.  

This may be good enough, but I personaly wanted to think about it more.  I suspect that if we went with a squelching approach as opposed to a publishing approach, more than just these emails should be squelched.  Again, I just haven't had the time to think it through, but that is what I am thinking so far.

Patch-set: 6