Update patch set 1
Patch Set 1:

One of the major problems I am stuck on right now is how to handle the data for patch sets of a change.

Originally I had thought of making the patch set data each its own document in the Lucene index. A query might match some of the patch set documents, and higher up in LuceneIndex.scan() we would try to convert a patch set document result into a change result. Unfortunately this doesn't work. A predicate like "owner:sop@google.com message:fixme" can only be true if both terms appear in the same document in Lucene.

There are some alternatives.

A brute force approach would be to index all of the patch set data as part of the change document. This provides simple query code, as the terms are all matching against a single document. Its also a fairly efficient query, Lucene just has to find documents where the terms are true. Its inefficient to update; making any modifications to a change require reindexing the entire change.

The other approach is to use a filter. For any operators in a query that apply to a patch set rather than a change, execute a recursive query on the Lucene index for patch set documents that match the relevant terms. Collect those documents, extract a field that contains the Change.Id, then dump these into a TermsFilter similar to what starredby() does with any() in LuceneQueryBuilder. This makes updates more incremental, as we don't need to reindex all change data all of the time. However it slows down searching because we have to translate the patch set's document object to the change's document object. This isn't done incrementally in Lucene, its sort of "all or nothing", and is why VisibleFilter has so much hackery to try and hoist the visiblity checking into the top level scan() loop, rather than down within the query evaluation logic.

A final alternative is to discard Lucene as the backend storage system and use something else that can handle huge volumes of data, like HBase. Build the inverted index with row keys of: <term, change-id>. Use Lucene's query support, but map change-id to docid in Lucene. This allows us to insert a bunch of different terms onto the same Lucene "document" out-of-order, without causing a full reindex of the document. Its also the most amount of developer work to implement. :-(

Patch-set: 1
Reviewer: Gerrit User 1001240 <1001240@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=0
Label: Verified=0