Update patch set 6
Patch Set 6:

I think the code looks good, but I would like more input from others if they like this feature, or think it is headed in the right direction, I am not quite convinced that it is.  I won't block this, but I don't really feel inclined to submit it either (yet).

To better explain what I was proposing in my emails, I proposed 2 buttons:

1) I require explicit patch publishing
2) Publish drafts comments when I publish patches

Now, when I said that I felt that this patch was "not what I had in mind in my proposal", I was not referring to the fact that it does not implement #2.  I felt that it did not implement #1!

What I envisioned by #1, is a mechanism for a user to upload a patch and have it remain in Gerrit in an unpublished state.  This state would mean that it would not be considered the latest patchset for a change by Gerrit.  This would not only imply that it would not send out an email to reviewers (*1), but that it would also not prevent submitting the previous patchset, it would not be visible to others when browsing the change, it would be deletable without a trace (just as draft comments are today), it would not be visible as a "needed-by" by other changes (its dependencies), if other changes depended on this unpublished patchset, they would also have to remain unpublished until this one is published (*2)...

In this unpublished state, I envisioned that the author could: see this latest patchset and compare it to other patchsets, make draft comments on this patchset, delete it, publish it.

On publishing, this patchset would become the latest for a change, and send out a published email.  This published email would include any draft comments on this patchset (this is not the same as #2 originally proposed which would
publish these draft comments plus draft comments from previous patchsets in one!)

*1 So, as you can see, my proposal is not so much a squelching of an email as it is an aggregation of emails.  A published patchset email would eventually be sent, it just may include comments.  Your patch never sends this email, and when a user receives comments on the new patchset, they may not even realize that a new patchset has been published, if they are not familiar with the patchset IDs (numbers), they may think the new comments simply apply to the old  patchset and may not realize that they should go look at the new code.

*2 I am not sure exactly how this would work, but it seems like it has some parallels to the way that changes are currently submitted when they depend on other unsubmitted changes...

One concern that I have with user preferences is whether they integrate well into a user's workflow.  A user preference seems appropriate when it really is a preference kind of thing, and perhaps for some, your implementation  would be, but for me I don't feel that it would be.  Let me explain:  for certain patchsets I might want this workflow because I expect to add comments to it.  For other patchsets however, I don't rely plan on adding comments to them, so I would want the old workflow so that it send out an email for me.  A user preference does not seem like an easy way to switch between these two different approaches regularly.  It suffers from the same problem the current "update" button does with patchset preferences (I want to see the whole file now, but I want my prefs to normally show 25 lines).  This has bothered me so much, that I even just uploaded a change  today to provide an "apply" button:
https://review.source.android.com/#/c/24356/
However, I am not sure that an apply would work at all here, but the proposed workflow above would likely be very similar to having an "apply" button.

Patch-set: 6