Update patch set 2
Patch Set 2:

Removing 'Patch Set' from  the column headers reduces the width of the columns only if there are no comments for the file in the patch set. If there are comments we still display 'X comments'. Not sure, but maybe for the comment row we should also show only show the number and skip the 'comments' string? What do you think?

Looking at the History table I now also think that it would look nicer if the patch set numbers in the first row would be in the center of the cell, rather than being right aligned.

Patch-set: 2