Update patch set 1
Patch Set 1: (2 inline comments)

I kind of understood that this was not meant as final version. 
You may include '[RFC]' in the beginning of your commit message to mark this change as 'request for comments'. I think this is common practice for such cases.

I think I and also Martin expressed in the summary comment that the idea of having this tutorial is very welcome. 

To me the structure of this document looks reasonable, as well as its content. I actually like that it repeats some infos from the other documents. Having a selected set of infos in one place, makes it easier for new users to get started. I also like that it concentrates on the straight-forward case and leaves out most of the corner cases. Regarding the command line output, I as a user always like when a concrete example of the command with its output is given. So as it is, it already looks good to me too.

Patch-set: 1