Update patch set 1
Patch Set 1:

What exactly do you mean by subtype? AFAIK the problem is that on client side we don't get the original exception but only a RemoteJsonException. From the RemoteJsonException we can only conclude to the original exception by checking the message.

What would be possible to avoid different exception classes is to throw always the same exception but with different messages, would this be better?

Patch-set: 1