Update patch set 2
Patch Set 2:

Martin and I discussed his scenario a bit offline. I think it makes the most sense for this change to go in as-is (workflow-wise), then once Martin gets the change-branch features merged we could add a follow-up change here that prompts you to select a valid branch if you try to restore a change where the destination branch does not exist. Sound reasonable?

Patch-set: 2