Update patch set 2
Patch Set 2:

If you read this thread, I think you will find that Gerrit is still missing critical functionality before your change would be a good idea:
http://code.google.com/p/gerrit/issues/detail?id=349&sort=-stars&colspec=ID%20Type%20Stars%20Milestone%20Status%20Priority%20Owner%20Summary

Specifically, Gerrit will need to be able to store change data in project repos so that change data can be removed/readded with project repos.  

So, while your change is a bit premature, if it sits here until the rest of the functionality is implemented, it should likely be considered for submission at that point.

As for the gitweb/cgit problem, that seems like a gap in the project state functionality and should likely be fixed by a change which targets applying the states filtering to gitweb, I think that fix is desirable with or without your archiving project change.

Patch-set: 2