Update patch set 4
Patch Set 4:

I agree with your comment about having new column in the UI branches table. We should push new patchset with it.

About the UI looking, we agree it is more similar to the 2.1.8 version of access control and the 2.2.2 version of UI branches table. But it does aggregate the undo functionality. Let me see if I understood it correctly: is your suggestion to modify it to have the exactly appearance than the 2.2.2 UI of access controls or would you agree on having this appearance providing the edition of an already added submit action and without an edition panel in the bottom?

Patch-set: 4
Reviewer: Gerrit User 1003770 <1003770@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=0