Update patch set 1
Patch Set 1:

This is cool.  Perhaps another useful role might be something like a release planner, someone with the ability to create and delete branches and create tags.  They probably also need to be able to push merge comits.

After that, perhaps a project lead, with owner on the project, and the ability to delegate branch ownership.  I am not sure what else.

Patch-set: 1