Update patch set 2
Patch Set 2:

This is really tough from a UX perspective. I completely agree that the new text is better but I'm not sure that its ideal.

I'd think 99% of people don't need to submit and would be better served if submission was just rules based. Then you just add your review and if the rules say go, it does.

Where I work, only leads have "Submit" and they only have it if "Gerritbot" is some how failing. Otherwise we just hide the submit concept all together because its just silly for us. If the code can go in, it should. If it can't, it shouldn't.

Patch-set: 2