Update patch set 5
Patch Set 5:

I am trying to understand the purpose of the ChangeCache class.

It looks to me that the top-level Map(Change.Id > ChangeCache) could also
be Map(Change.Id > ChangeDetailCache) and that for a given changeId there
is 1:1 relationship between the ChangeCache and the ChangeDetailCache instances for this changeId.

Understanding the intention with the ChangeCache would make the code review easier.

Patch-set: 5
Label: Code-Review=0
Label: Verified=0