Update patch set 1
Patch Set 1:

Would be great to have a blocking push for some cases, for instance if the project has a replication, don't finish the push until the replication to all the hosts are complete.

This should probably have a timeout and tell the user that replication failed or succeeded for the host in question in the output.

Patch-set: 1
Label: Code-Review=0