Update patch set 3
Patch Set 3:

> Of course. But it is a long process and certainly not possible in all corporate environments. Regardless, I think the excessive querying should be cached and patched in Gerrit (where it belongs).

Well, until now changes have not been cached, probably because they have critical data.  Most caches that Gerrit has is for non critical stuff.
> Afaik, you cannot change destination-branches today. When that functionality exists, or is needed (like for "topic:moveChange"), I hope/expects a suitable listener interface will be introduced.

But how would this listener interface update your slaves?  This is the same problem that master master would have to solve.


> I also think that changing destination branch would/should generate a new patchset, so you could follow the change history and have access to the older commit graphs. That would also make sure that caches like this up-to-date.

Changing the destination branch does not actually change the commit.  It just repositions it to be submited to a different branch, so there will be no new patchset or graph changes.

Patch-set: 3