Update patch set 3
Patch Set 3:

> Well, until now changes have not been cached, probably because they have critical data. Most caches that Gerrit has is for non critical stuff. 

Or as Shawn said when I asked the same thing: "Because I'm lazy" :)

> "Most caches that Gerrit has is for non critical stuff."

I'm not sure this makes complete sense to me, since I'm not sure how you rate what is "Non critical stuff" and how you rate the TagCache, GroupsCache, PermissionCache, SSHKeys etc against this one. I think you need to clarify, if you want me to meet this argument in a fair way.

> But how would this listener interface update your slaves? This is the same problem that master master would have to solve.

See below.

> Changing the destination branch does not actually change the commit. It just repositions it to be submited to a different branch, so there will be no new patchset or graph changes.

Ok. I can see that I need to elaborate a bit. I figured you absolutely wanted to tie the destination ref to the patch set (instead of the change) and create a new PatchSet when the destination-branch changes. Several reasons:

* The change of destination should be trace-able.

* You likely want to do a rebase (at least have the possibility to do one) against the new destination branch.

* Since the resulting (post submit) graph will certainly change, the reviews should not be kept. Different branches have different workflow and processes tied in. Hence, only the vetoes should be copied (or copied accordingly to active filter plug-ins). Therefore the same rules should be applied in this scenario, as whenever an old PatchSet becomes obsolete. So just giving it a new patch would absolutely make sense here.

I can also imagine a lot of complex Build/CI-environments getting issues if the destination branch changes, but not seeing a new patchset in the stream-events.

To be honest I have not had time to look into the changeTopic branch. So perhaps my assumption is way out there. If so, I apologies.

I'm not sure how the "Change destination branch" feature could be implemented in a really secure and reliable way, without taking scenarios like these into account. And if it would approach the changing of destination branch this way, the issues with the stale change cache goes away. If not, perhaps it should be possible to opt-out the changeBranch functionality.

---

Since I do not really agree with some of the comments (which I still thank you greatly for and appreciate very much) I would really love to see a few more core devs here, so we can come to conclusion if this feature ever will be integrated. And if not, I chould spend my time doing more productive things.

Patch-set: 3