Update patch set 1
Patch Set 1:

I hadn't thought of showing the blue breaks, but yeah that is a good idea.  I'll see if I can add those later today or tomorrow depending on my schedule.

As for showing the overview bar inside the scrollbar - we could try that, but it would be confusing.  The bar doesn't scroll with the rest of the page, so it shouldn't be inside the scrollable area (imo).  I think Eclipse got this one right.  Are there other overview bars we could look at for inspiration?

Patch-set: 1