Update patch set 1
Patch Set 1: I would prefer that you didn't submit this; Fails

When a change is abandoned the 'Can Merge' is not shown since the change is now closed. This is good, but when the change is restored again, 'Can Merge' stays hidden, although it should now be displayed again.

Patch-set: 1
Label: Code-Review=-1
Label: Verified=-1