Update patch set 1
Patch Set 1: (3 inline comments)

Thanks Sasa!
I like your dry idea! I name it 'All' now.
The above file name is a bit far from the diff table thus it would be not obvious ... and it is not clear for old version or new version, so I did not let it clickable.

Patch-set: 1