Update patch set 2
Patch Set 2:

Hi Sasa
I think the file-name is a bit far from the diff table thus it would be not obvious to user to find they can double click the file-name to add file comments. but the really problem for me is I cannot figure out which version we should add the file comments after user double click the file-name. maybe it should be on the new version. if you agree on it I will add a new pate set accordingly .

welcome other cores' idea.
/Bruce

Patch-set: 2