Update patch set 3
Patch Set 3:

This change (well along with the others in the series but I think its this one that had the most impact) has dropped some git push times on repositories with a long history of changes from 89s... to 13s on one of our servers. Its a huge difference.

Patch-set: 3
Label: Code-Review=0
Label: Verified=0