Update patch set 5
Patch Set 5:

"For # 3 I suggest just allowing a user to double click anywhere in the header 
(not just on the "Patch Set" words) to begin adding a comment."

 
I find this need a lot of refactoring work, and I have not
get any positive response from other cores by far, so I
suggest in #3 still keep double click the "Patch Set" words
 and carry "double click anywhere in the header " out in a 
later separate change when ALL cores agree, let's call it 
as #5, But it is outside of the scope of the feature. I 
like #5 because I like to add the middle border line, also that would make header look better in IE, chrome that it is currently.

Patch-set: 5